Add in auth parent.
Since the app now requires a login, we need to prevent the user from
doing anything until they login. I've added a parent view controller to
avoid breaking anything inside the main "app" and still add this
ability. The parent has three possible states, determining the user's
status, logged in, or not logged in. It will depend on main storyboard
for the logged in status to make the app work like normal. For the case
where the user is not logged in, it will look for a new storyboard which
contains the auth flow. While determining this status, the app will show
an adorable Pokemon and an activity indicator.