Added in some basic API support.
I'm not going to interface with an actual server, but for all pieces
we've built so far, it doesn't matter. This could just be a user saved
locally, or in this case faked to always exist. This lets us see what
the app does once it thinks that the user is logged in.