Add the ability to logout and login flow.
The app pretends to be logged in initially, but you need to be able to
logout to see how the login flow works too in such an arrangement. The
second tab now has a logout button. This will trigger the auth parent to
show the AuthFlow's controllers instead of the main content. This will
dump the memory used by anything in the main content because it's no
longer a part of the hierarchy. Actual logout may involve some logic or
talking to a server, but most of the app doesn't care, that's up for the
API to decide. Login is done as individual steps just to artificially
inflate the number of steps needed to show how login/signup can work
with multiple steps.

Logging in with email and password in this example arbitrarily decides
the info is valid, but adding any logic to say, check with an actual
server would happen in the AuthFlow, the controllers wouldn't
necessarily need to be modified.

Also, signup isn't possible right now because that comes later.