Better fake login.
This lifts the user manager up one level so that login can actually
pretend to check against whatever database/api the user manager
controls. This makes it a bit more realistic.

Once login is successful (it's fake now, so it's easy to succeed), the
auth parent goes back to showing the main interface and everyone is
happy. This completes the loop and covers everything except new user
registration, which would require making use of the other stages in the
AuthFlow.